Move common.css into layout folder
