+++ src/components/layout/common.css
@@ -1,5 +1,5 @@
-.form-control{
-  padding: .5rem;
+.form-control {
+  padding: 0.5rem;
 }
 
 .form-control label {
